py3Dmol.view now expects ints for width and height. (Blank plot if not.) [closes #412]

+++ python/BioSimSpace/Sandpit/Exscientia/Align/_align.py
@@ -1627,11 +1627,11 @@ def viewMapping(
         viewergrid = (1, 2)
         viewer1 = (0, 1)
         width = pixels
-        height = pixels / 2
+        height = int(pixels / 2)
     else:
         viewergrid = (2, 1)
         viewer1 = (1, 0)
-        width = pixels / 2
+        width = int(pixels / 2)
         height = pixels
 
     # Create the view.
